[modified] Використано метод get_jwt_identity(), захищено ендпоїнт із видаленням користувача

+++ project/app.py
@@ -1,6 +1,6 @@
 import os
 from flask import Flask, request, jsonify
-from flask_jwt_extended import JWTManager, create_access_token, jwt_required
+from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity
 from flask_migrate import Migrate
 from flask_restful import Api
 from marshmallow import Schema, fields, validate, ValidationError
@@ -9,7 +9,7 @@ from passlib.handlers.pbkdf2 import pbkdf2_sha256
 
 app = Flask(__name__)
 app.config[
-     'SQLALCHEMY_DATABASE_URI'] = 'postgresql://postgres:12345@localhost:5432/Lab4'
+    'SQLALCHEMY_DATABASE_URI'] = 'postgresql://lab_4_db_user:yWLIJiZE6CIgdHNSeo1cRkYL2fNHPBGL@dpg-cm3fj2mn7f5s73bo8fi0-a.oregon-postgres.render.com/lab_4_db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 os.environ["JWT_SECRET_KEY"] = "283129114377609256790918969598140128636"
 app.config["JWT_SECRET_KEY"] = os.getenv("JWT_SECRET_KEY")
@@ -90,6 +90,7 @@ def register_user():
 
     return jsonify({'message': 'User created!'}), 201
 
+
 # Логін користувача
 @app.route('/login', methods=['POST'])
 def login_user():
@@ -104,6 +105,7 @@ def login_user():
 
     access_token = create_access_token(identity=user.user_id)
     return jsonify(access_token=access_token), 200
+
 
 # Пользователи
 @app.route('/user', methods=['POST'])
@@ -133,7 +135,12 @@ def get_user(user_id):
 
 
 @app.route('/user/<user_id>', methods=['DELETE'])
+@jwt_required()
 def delete_user(user_id):
+    current_user_id = get_jwt_identity()
+    if current_user_id != user_id:
+        return jsonify({'error': 'You are not authorized to delete this user'}), 403
+
     user = User.query.get(user_id)
     if user:
         db.session.delete(user)
@@ -142,19 +149,24 @@ def delete_user(user_id):
     return jsonify({'error': 'User not found'}), 404
 
 
-# Категории
+# Категорії
+# Створено перевірку, чи співпадає ID поточного користувача із вказаним параметром owner_id
 @app.route('/category', methods=['POST'])
 @jwt_required()
 def create_category():
     try:
         category_data = category_schema.load(request.get_json())
+        current_user_id = get_jwt_identity()
         new_category = Category(category_id=category_data['category_id'],
                                 category_name=category_data['category_name'],
                                 visibility=category_data.get('visibility'),
                                 owner_id=category_data.get('owner_id'))
+
+        if new_category.owner_id and new_category.owner_id != current_user_id:
+            return jsonify({'error': 'You are not allowed to set a different owner for the category'}), 403
+
         db.session.add(new_category)
         db.session.commit()
-
         return jsonify(category_schema.dump(new_category))
     except ValidationError as err:
         return jsonify({'error': err.messages}), 400
@@ -181,11 +193,15 @@ def get_category(category_id):
     return jsonify({'error': 'Category not found'}), 404
 
 
+# Створено перевірку, чи співпадає id поточного користувача із id власника категорії
 @app.route('/category/<category_id>', methods=['DELETE'])
 @jwt_required()
 def delete_category(category_id):
     category = Category.query.get(category_id)
+    current_user_id = get_jwt_identity()
     if category:
+        if category.owner_id != current_user_id:
+            return jsonify({'error': 'You are not authorized to delete this category'}), 403
         db.session.delete(category)
         db.session.commit()
         return jsonify({'message': 'Category deleted successfully'})
@@ -193,6 +209,7 @@ def delete_category(category_id):
 
 
 # Записи
+# Створено перевірку, вказане при створенні категорії user_id повинне співпадати із ID поточного користувача
 @app.route('/record', methods=['POST'])
 @jwt_required()
 def create_record():
@@ -200,28 +217,26 @@ def create_record():
         record_data = record_schema.load(request.get_json())
     except ValidationError as err:
         return jsonify({'error': err.messages}), 400
-
+    current_user_id = get_jwt_identity()
     category_id = record_data['category_id']
     category = Category.query.get(category_id)
     if category is None:
         return jsonify({'error': 'Category not found'}), 404
-
     visibility = category.visibility
     if visibility == 'private':
         user_id = record_data['user_id']
-        if user_id != category.owner_id:
-            return jsonify({'error': 'You are not allowed to use this category'}), 403
 
+        if user_id != current_user_id:
+            return jsonify({'error': 'You are not allowed to use this category'}), 403
     new_record = Record(id=record_data['id'],
                         user_id=record_data['user_id'],
                         category_id=record_data['category_id'],
                         creation_data=record_data['creation_data'],
                         cost=record_data['cost'])
-
     db.session.add(new_record)
     db.session.commit()
-
     return jsonify(record_schema.dump(new_record))
+
 
 
 @app.route('/records', methods=['GET'])
@@ -242,11 +257,15 @@ def get_records():
     return jsonify(record_schema.dump(records, many=True))
 
 
+# Створено перевірку, тепер запис може тільки той користувач, який його створив
 @app.route('/record/<record_id>', methods=['DELETE'])
 @jwt_required()
 def delete_record(record_id):
     record = Record.query.get(record_id)
+    current_user_id = get_jwt_identity()
     if record:
+        if record.user_id != current_user_id:
+            return jsonify({'error': 'You are not authorized to delete this record'}), 403
         db.session.delete(record)
         db.session.commit()
         return jsonify({'message': 'Record deleted successfully'})
